backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
asyncpg>=0.30
pydantic>=2.0
bcrypt>=4.0
PyJWT>=2.8
httpx>=0.27
